@@ -14,12 +14,20 @@
         "source": "https://github.com/bowphp/app"
     },
     "require": {
-        "php": "^7.1|^8.0",
-        "bowphp/framework": "dev-master",
-        "bowphp/policier": "^2.1"
+        "php": "^8.1",
+        "bowphp/framework": "5.x-dev",
+        "bowphp/policier": "^2.1",
+        "ext-openssl": "*",
+        "ext-pdo": "*",
+        "ext-mbstring": "*",
+        "ext-xml": "*",
+        "ext-json": "*",
+        "ext-pdo_mysql": "*",
+        "ext-intl": "*",
+        "ext-pdo_sqlite": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7",
+        "phpunit/phpunit": "^8",
         "monolog/monolog": "^1.22",
         "squizlabs/php_codesniffer": "3.*"
     },
@@ -39,7 +47,7 @@
         ]
     },
     "scripts": {
-        "code-formating": "phpcbf --standard=psr2 --tab-width=4 app tests config migrations seeders routes",
+        "lint": "phpcbf --standard=psr12 --tab-width=4 app tests config migrations seeders routes",
         "test": "phpunit --configuration phpunit.dist.xml",
         "post-root-package-install": [
             "@php -r \"file_exists('.env.json') || copy('.env.example.json', '.env.json');\""
@@ -54,7 +62,7 @@
         "optimize-autoloader": true
     },
     "suggest": {
-        "twig/twig": "For the use of template engine 'Twig'",
-        "papac/bow-blade": "For the use of template engine 'Blade'"
+        "twig/twig": "For using Twig template engine",
+        "papac/bow-blade": "For using Blade template engine"
     }
 }
