@@ -7,11 +7,23 @@ use Bow\Configuration\Loader as ApplicationLoader;
 class Kernel extends ApplicationLoader
 {
     /**
-     * Get app namespace
+     * Define your events
      *
      * @return array
      */
-    public function namespaces()
+    public function events(): array
+    {
+        return [
+            // Put your event here
+        ];
+    }
+
+    /**
+     * Define the app namespace
+     *
+     * @return array
+     */
+    public function namespaces(): array
     {
         return [
             'controller' => 'App\\Controllers',
@@ -21,6 +33,7 @@ class Kernel extends ApplicationLoader
             'model' => 'App\\Models',
             'service' => 'App\\Services',
             'event' => 'App\\Events',
+            'listener' => 'App\\Listeners',
             'exception' => 'App\\Exceptions',
             'producer' => 'App\\Producers',
             'command' => 'App\\Commands',
@@ -28,11 +41,11 @@ class Kernel extends ApplicationLoader
     }
 
     /**
-     * The middleware lists
+     * Define the app middlewares
      *
      * @return array
      */
-    public function middlewares()
+    public function middlewares(): array
     {
         return [
             'csrf' => \App\Middlewares\RequestCsrfMiddleware::class,
@@ -46,7 +59,7 @@ class Kernel extends ApplicationLoader
      *
      * @return array
      */
-    public function configurations()
+    public function configurations(): array
     {
         return [
             /**
@@ -70,16 +83,23 @@ class Kernel extends ApplicationLoader
              * Add your Custom Settings here.
              */
             // \Policier\Bow\PolicierConfiguration::class,
+
+            /**
+             * Should be the last to loading
+             */
+            \App\Configurations\ApplicationConfiguration::class,
         ];
     }
 
     /**
      * Service Bootstrap
      *
-     * @return void
+     * @return ApplicationLoader
      */
-    public function boot()
+    public function boot(): ApplicationLoader
     {
         parent::boot();
+
+        return $this;
     }
 }
