@@ -1,7 +1,7 @@
 <?php
 
 use Bow\Console\Color;
-use Bow\Console\ArgOption as Argument;
+use Bow\Console\Argument;
 
 $console->addCommand('hello', function (Argument $argument) {
     echo Color::green("hello, bow task runner.");
